refactor(app): separate deposit flows in pages
