+++ apps/next/pages/deposit/index.tsx
@@ -1,9 +1,9 @@
 import { DepositScreen } from 'app/features/deposit/screen'
 import { HomeLayout } from 'app/features/home/layout.web'
+import { TopNav } from 'app/components/TopNav'
 import Head from 'next/head'
 import { userProtectedGetSSP } from 'utils/userProtected'
 import type { NextPageWithLayout } from '../_app'
-import { TopNav } from 'app/components/TopNav'
 
 export const Page: NextPageWithLayout = () => {
   return (
